pyproject.toml
[tool.poetry]
name = "logistic-service"
version = "0.1.0"
description = ""
authors = ["Alex <124603897+OleksandrBrusyltsev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"

black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

psycopg2-binary = "^2.9.9"
django-redis = "^5.4.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exlude = '''
    venv/
    '''

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120


[tool.flake8]
exclude = ".git,__pychache__,venv,__init__.py"
max-line-length = 119
ignore = ["E203", "E266", "E501", "W503"]
